Move provider dashboard screen state to dedicated file
